# shared/openapi.json
{
  "components": {
    "schemas": {
      "AgentStatus": {
        "enum": [
          "down",
          "up"
        ],
        "title": "AgentStatus",
        "type": "string"
      },
...
      "AuthorizationCredentials": {
        "properties": {
          "login_session_token": {
            "title": "Login Session Token",
            "type": "string"
          },
          "login_session_uid": {
            "format": "uuid",
            "title": "Login Session Uid",
            "type": "string"
          }
        },
        "required": [
          "login_session_uid",
          "login_session_token"
        ],
        "title": "AuthorizationCredentials",
        "type": "object"
      },
...
      "CreateAgentDTO": {
        "properties": {
          "ip": {
            "format": "ipvanyaddress",
            "title": "Ip",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "port": {
            "title": "Port",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "ip",
          "port"
        ],
        "title": "CreateAgentDTO",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LoginCredentials": {
        "properties": {
          "password": {
            "title": "Password",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginCredentials",
        "type": "object"
      },
      "MyAgentDTO": {
        "properties": {
          "compose_file": {
            "readOnly": true,
            "title": "Compose File",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "ip": {
            "format": "ipvanyaddress",
            "title": "Ip",
            "type": "string"
          },
          "is_suspended": {
            "title": "Is Suspended",
            "type": "boolean"
          },
          "keypair": {
            "$ref": "#/components/schemas/MyKeypairDTO"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "port": {
            "title": "Port",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/AgentStatus"
          }
        },
        "required": [
          "id",
          "name",
          "ip",
          "port",
          "status",
          "is_suspended",
          "created_at",
          "keypair",
          "compose_file"
        ],
        "title": "MyAgentDTO",
        "type": "object"
      },
      "MyKeypairDTO": {
        "properties": {
          "public_key_openssh": {
            "title": "Public Key Openssh",
            "type": "string"
          }
        },
        "required": [
          "public_key_openssh"
        ],
        "title": "MyKeypairDTO",
        "type": "object"
      },
...
      "Register": {
        "properties": {
          "password": {
            "title": "Password",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "Register",
        "type": "object"
      },
...
      "UpdateAgentDTO": {
        "properties": {
          "ip": {
            "format": "ipvanyaddress",
            "title": "Ip",
            "type": "string"
          },
          "is_suspended": {
            "title": "Is Suspended",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "port": {
            "title": "Port",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "ip",
          "port",
          "is_suspended"
        ],
        "title": "UpdateAgentDTO",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/agents": {
      "get": {
        "operationId": "get_my_agents_agents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MyAgentDTO"
                  },
                  "title": "Response Get My Agents Agents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get My Agents"
      },
      "post": {
        "operationId": "create_agent_agents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyAgentDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Agent"
      }
    },
    "/agents/{agent_id}": {
      "delete": {
        "operationId": "delete_agent_agents__agent_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Agent"
      },
      "put": {
        "operationId": "update_agent_agents__agent_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAgentDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyAgentDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Agent"
      }
    },
...
    "/login": {
      "post": {
        "operationId": "login_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCredentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizationCredentials"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login"
      }
    },
    "/register": {
      "post": {
        "operationId": "register_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Register"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register"
      }
    }
  }
}